Refactor: create `combineValidations` utility & unify validators
[Now all `validate...` functions return an array of validation errors.
422 errors are thrown inside resolver functions.]
