+++ ex-4-graphql/server/graphql/mutations/post/createPost.js
@@ -1,5 +1,6 @@
 const { GraphQLString } = require('graphql');
 
+const ErrorHandler = require('../../../utils/ErrorHandler');
 const { Post: PostModel } = require('../../../models');
 const { Post: PostType } = require('../../types');
 const { post: validatePost } = require('../../../validators/feed');
@@ -12,7 +13,11 @@ const createPost = {
 		image: { type: GraphQLString }
 	},
 	resolve: async function (source, args) {
-		await validatePost(args);
+		const errors = await validatePost(args);
+
+		if (errors.length > 0) {
+			throw new ErrorHandler(422, 'Post validation failed.', errors);
+		}
 
 		/* temporarily fixed image and creator to avoid GraphQL errors */
 		const post = await PostModel.create({
